+++ src/config.py
@@ -40,7 +40,7 @@ from src.llm import generation_params as llm_generation_params
 logger = logging.getLogger(__name__)
 
 DEFAULT_ALPHASIFT_INSTALL_SPEC = (
-    "git+https://github.com/CXTretar/alphasift.git@ea1360787dcbeffc081887730fbb57f74f83dd85"
+    "git+https://github.com/CXTretar/alphasift.git@5d8ba5acd8ffc0c12ba8ffcac937f7af3984fa97"
 )